% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bistable_circuitGenes.R
\name{get_bistable_circuitGenes}
\alias{get_bistable_circuitGenes}
\title{get_bistable_circuitGenes}
\usage{
get_bistable_circuitGenes(data, Switch_TF, Transient_TF)
}
\arguments{
\item{data}{refer to time series of committed data}

\item{Switch_TF}{Pro-screening candidate genes of Switch_TF}

\item{Transient_TF}{Pro-screening candidate genes of Transient_TF}
}
\value{
bistable circuitGenes
}
\description{
function to get bistable circuit genes
}
