% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diverGenes.R
\name{get_diverGenes}
\alias{get_diverGenes}
\title{get_diverGenes}
\usage{
get_diverGenes(ODE_pair_results)
}
\arguments{
\item{ODE_pair_results}{refer to bistable circuitGenes}
}
\value{
diver genes
}
\description{
function to get diver genes
}
