% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_moduleAnalysis.R
\name{get_moduleAnalysis}
\alias{get_moduleAnalysis}
\title{get_moduleAnalysis}
\usage{
get_moduleAnalysis(data, PPI_gene)
}
\arguments{
\item{data}{refer to time series of committed data}

\item{PPI_gene}{refer to bistable circuitGenes}
}
\value{
gene modules Analysis
}
\description{
function to get module Analysis
}
