% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_moduleAnalysis_A.R
\name{get_moduleAnalysis_A}
\alias{get_moduleAnalysis_A}
\title{modules_identification}
\usage{
get_moduleAnalysis_A(data, Switch_TF, Transient_TF)
}
\arguments{
\item{data}{refer to time series of committed data}

\item{Switch_TF}{Pro-screening candidate genes of Switch_TF}

\item{Transient_TF}{Pro-screening candidate genes of Transient_TF}
}
\value{
gene modules
}
\description{
function to get modules identification
}
