% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_identification.R
\name{modules_identification}
\alias{modules_identification}
\title{modules_identification}
\usage{
modules_identification(data, ODE_single_results)
}
\arguments{
\item{data}{refer to time series of committed data}

\item{ODE_pair_results}{refer to bistable circuitGenes}
}
\value{
gene modules
}
\description{
function to get modules identification
}
