% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_identification_A.R
\name{modules_identification_A}
\alias{modules_identification_A}
\title{modules_identification}
\usage{
modules_identification_A(data)
}
\arguments{
\item{data}{refer to time series of committed data}
}
\value{
gene modules
}
\description{
function to get modules identification
}
