% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Screen_TF.R
\name{Switch_nls}
\alias{Switch_nls}
\title{screen switch-like and transient TFs}
\usage{
Switch_nls(scExp)
}
\arguments{
\item{scExp}{refer to time series of committed data}
}
\value{
Pro-screening candidate genes
}
\description{
function to Pro-screening candidate genes
}
